+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Dev
+- Use "compilation" and "afterProcessAssets" hooks instead of "afterEmit"
+- Use "emitAsset" to generate map file
+
 # Version 0.2.1
 - Various configuration adjustments
 - Add README file